requirements.txt
openai
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-anthropic
langsmith
langgraph
pillow
numpy
tqdm
requests
tavily-python
pandas
matplotlib
seaborn
jupyter